Refactor : 강의 관리 Controller 메소드 위치변경
/info/mycourselist 도메인 위치변경 및 메소드 위치변경
