+++ src/main/resources/templates/course/mycourselist.mustache
