+++ src/main/resources/templates/courseboard/courseInfoUpdate.mustache
@@ -59,7 +59,7 @@
                     </div>
 
                     <hr class="my-4">
-                    <button id="courseUpdateBtn" class="w-100 btn btn-primary btn-lg">강의 업데이트 하기</button>
+                    <button id="courseUpdateBtn" type="button" class="w-100 btn btn-primary btn-lg">강의 업데이트 하기</button>
                 </form>
             </div>
         </div>
